+++ docs/conf.py
@@ -12,6 +12,7 @@
 # serve to show the default.
 
 import sys, os
+import datetime
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the
@@ -42,7 +43,8 @@ master_doc = 'index'
 
 # General information about the project.
 project = u'taxtastic'
-copyright = u'2011-2013, Noah Hoffman, Erick Matsen, Brian Hodges'
+copyright = u'2011-{}, Noah Hoffman, Erick Matsen, Brian Hodges, Connor McCoy'.format(
+    datetime.date.today().strftime('%Y'))
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the