+++ taxtastic/refpkg.py
@@ -20,27 +20,35 @@ Note that Refpkg objects are *NOT* thread safe!
 #
 #    You should have received a copy of the GNU General Public License
 #    along with taxtastic.  If not, see <http://www.gnu.org/licenses/>.
+import errno
+import os
+import time
+import warnings
+
+import Bio.Phylo
+import Bio.SeqIO
 import contextlib
-from decorator import decorator
+import copy
+import csv
+import functools
+import hashlib
+import json
+import shutil
 import subprocess
 import tempfile
-import hashlib
-import shutil
-import os
-import copy
-import json
-import time
-import csv
-import errno
-import warnings
 import zipfile
 
-import Bio.SeqIO
-import Bio.Phylo
+from decorator import decorator
 
 from taxtastic import utils, taxdb
 
+
 FORMAT_VERSION = '1.1'
+
+
+class DerivedFileNotUpdatedWarning(UserWarning):
+    pass
+
 
 def md5file(fobj):
     md5 = hashlib.md5()
@@ -53,21 +61,20 @@ def md5file(fobj):
 def scratch_file(unlink=True, **kwargs):
     """Create a temporary file and return its name.
 
-    Additional arguments are passed to ``tempfile.mkstemp``
+    Additional arguments are passed to :class:`tempfile.NamedTemporaryFile`
 
     At the start of the with block a secure, temporary file is created
     and its name returned.  At the end of the with block it is
     deleted.
     """
+    kwargs['delete'] = False
+    tf = tempfile.NamedTemporaryFile(**kwargs)
+    tf.close()
     try:
-        tmp_fd, tmp_name = tempfile.mkstemp(text=True, **kwargs)
-        os.close(tmp_fd)
-        yield tmp_name
-    except ValueError:
-        raise
-    else:
+        yield tf.name
+    finally:
         if unlink:
-            os.unlink(tmp_name)
+            os.unlink(tf.name)
 
 
 def manifest_template():
@@ -448,9 +455,10 @@ class Refpkg(object):
         self._add_file(key, new_path)
         md5_value = md5file(open(new_path))
         self.contents['md5'][key] = md5_value
-        self._log('Updated file: %s=%s' % (key,new_path))
-        if key == 'tree_stats':
-            self.update_phylo_model(None, new_path)
+        self._log('Updated file: %s=%s' % (key, new_path))
+        if key == 'tree_stats' and old_path:
+            warnings.warn('Updating tree_stats, but not phylo_model.',
+                          DerivedFileNotUpdatedWarning, stacklevel=2)
         return old_path
 
     @transaction
@@ -474,7 +482,7 @@ class Refpkg(object):
                 self.update_file('tree', name)
         self._log('Rerooting refpkg')
 
-    def update_phylo_model(self, stats_type, stats_file):
+    def update_phylo_model(self, stats_type, stats_file, frequency_type=None):
         """Parse a stats log and use it to update ``phylo_model``.
 
         ``pplacer`` expects its input to include the deatils of the
@@ -487,7 +495,20 @@ class Refpkg(object):
         parameter must be 'RAxML', 'PhyML' or 'FastTree', depending on which
         program generated the log. It may also be None to attempt to guess
         which program generated the log.
+
+        :param stats_type: Statistics file type. One of 'RAxML', 'FastTree', 'PhyML'
+        :param stats_file: path to statistics/log file
+        :param frequency_type: For ``stats_type == 'PhyML'``, amino acid
+         alignments only: was the alignment inferred with ``model`` or
+         ``empirical`` frequencies?
+
         """
+
+        if frequency_type not in (None, 'model', 'empirical'):
+            raise ValueError('Unknown frequency type: "{0}"'.format(frequency_type))
+        if frequency_type and stats_type not in (None, 'PhyML'):
+            raise ValueError('Frequency type should only be specified for '
+                             'PhyML alignments.')
 
         if stats_type is None:
             with open(stats_file) as fobj:
@@ -511,7 +532,8 @@ class Refpkg(object):
         elif stats_type == 'FastTree':
             parser = utils.parse_fasttree
         elif stats_type == 'PhyML':
-            parser = utils.parse_phyml
+            parser = functools.partial(utils.parse_phyml,
+                                       frequency_type=frequency_type)
         else:
             raise ValueError('invalid log type: %r' % (stats_type,))